refactor: 피드백 내용 반영 1. 불필요한 equals, hashcode 오버라이딩 제거 2. while로 계속 남은 티켓오더 확인하는 로직 for문으로 변경 3. Lotto 번호 검증 로직 Lotto에 넣음
