+++ src/main/java/step3/model/lotto/Lotto.java
@@ -2,13 +2,17 @@ package step3.model.lotto;
 
 
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+import step3.model.value.ErrMsg;
+import step3.model.value.Rule;
 
 public class Lotto{
     private List<LottoNumber> lotto;
     public Lotto(List<Integer> lotto) {
+        verifyLotto(lotto);
         this.lotto =createLotto(lotto);
     }
 
@@ -22,27 +26,13 @@ public class Lotto{
     public int getMatchCount(Lotto winningLotto){
         return (int) lotto.stream().filter(winningLotto::isMatched).count();
     }
-    public boolean isAllMatch(Lotto inputLotto){
-        return this.lotto.stream().allMatch(inputLotto::isMatched);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
+    private void verifyLotto(List<Integer> lottoNumbers) {
+        if(lottoNumbers.size() != Rule.LOTTO_NUMBER_LENGTH){
+            throw new IllegalArgumentException(ErrMsg.WRONG_LENGTH);
         }
-        if (!(o instanceof Lotto)) {
-            return false;
+        if(lottoNumbers.size() != new HashSet<>(lottoNumbers).size()){
+            throw new IllegalArgumentException(ErrMsg.DUPLICATED_INPUT);
         }
-        Lotto that = (Lotto) o;
-        Collections.sort(this.lotto);
-        Collections.sort(that.lotto);
-        return this.lotto.equals(that.lotto);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(lotto);
     }
 
     @Override
