+++ src/main/java/step3/model/machine/LottoMachine.java
@@ -3,6 +3,7 @@ package step3.model.machine;
 import java.util.ArrayList;
 import java.util.List;
 import step3.model.lotto.Lotto;
+import step3.model.lotto.LottoList;
 import step3.model.value.Rule;
 
 
@@ -15,16 +16,17 @@ public class LottoMachine {
         this.manualGenerator = new LottoManualGenerator();
     }
 
-    public List<Lotto> issueAutoLottoList(Order order) {
+    public LottoList issueAutoLottoList(Order order) {
         List<Lotto> lottoList = new ArrayList<>();
-        int ticketCount=0;
-        while(order.leftToTicketing(ticketCount)){
-            ticketCount+=1;
+        int ticketCount = order.getAutoTicketCount();
+
+        for (int i=0; i< ticketCount; i++) {
             lottoList.add(new Lotto(autoGenerator.generateLottoAuto(Rule.LOTTO_NUMBER_LENGTH)));
         }
-        return lottoList;
+        return new LottoList(lottoList);
     }
-    public List<Integer> createWinningLotto(String input) {
-        return manualGenerator.createLotto(input);
+    public Lotto createWinningLotto(String input) {
+        List<Integer> lotto = manualGenerator.createLotto(input);
+        return new Lotto(lotto);
     }
 }
