+++ src/main/java/step3/model/machine/Order.java
@@ -8,9 +8,6 @@ public class Order {
         this.ticketCount = ticketCount;
     }
 
-    public boolean leftToTicketing(int ticketNumber) {
-        return ticketCount>ticketNumber;
-    }
     public long getTotalPrice(int ticketPrice){
         return ticketCount*ticketPrice;
     }
@@ -18,5 +15,9 @@ public class Order {
     @Override
     public String toString() {
         return ticketCount+"개를 구매했습니다.";
+    }
+
+    public int getAutoTicketCount() {
+        return ticketCount;
     }
 }
